+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Pending
 
+* [Chore] Fix MRAID Error Logging [#16439](https://github.com/Applifier/unity-ads-webview/pull/16439)
+* [Load V5] Update Allow List GameIDs [#16438](https://github.com/Applifier/unity-ads-webview/pull/16438)
 * [Release 3.4.8] Add Webview Release Support [#16431](https://github.com/Applifier/unity-ads-webview/pull/16431)
 * [Chore] Update Dependencies [#16430](https://github.com/Applifier/unity-ads-webview/pull/16430)